requirements.txt
# Core
pytz==2018.7
Django==2.1.3
django-configurations==2.1
gunicorn==19.9.0
newrelic==4.6.0.106

# Auth
django-allauth==0.38.0
django-rest-auth==0.9.3

# For the persistence stores
psycopg2-binary==2.7.6.1
dj-database-url==0.5.0

#Redis
redis==3.0.1
django-websocket-redis==0.5.2

# Model Tools
django-model-utils==3.1.2
django_unique_upload==0.2.1

# Rest apis
djangorestframework==3.9.0
Markdown==3.0.1
#django-filter==2.0.0
drf-schema-adapter==0.9.77
djangorestframework-jwt==1.11.0
djangorestframework-filters==0.10.2.post0
drf-writable-nested==0.5.1

# Developer Tools
ipdb==0.11
ipython==7.1.1
mkdocs==0.17.5
flake8==3.6.0
django-extensions==2.1.3
pre-commit==1.12.0
black==18.9b0
better-exceptions==0.2.1

# Testing
mock==2.0.0
factory-boy==2.11.1
django-nose==1.4.6
nose-progressive==1.5.1
coverage==4.5.2
pytest==3.8.2
pytest-django==3.4.3
git+git://github.com/mverteuil/pytest-ipdb.git@master#egg=pytest-ipdb
pytest-cov==2.6.0

# Static and Media Storage
django-storages==1.7.1
boto3==1.9.47

# Field history
django-field-history==0.7.0

# dateutil
python-dateutil==2.7.5

# Address / Phone fields
django-address==0.2.0
django-phone-field==1.6.0
django-phonenumber-field==2.1.0
phonenumbers==8.10.0

django-webpack-loader==0.6.0

black
pre-commit